cast to int when checking if the caller is querying the actual value
instead of setting a new one (by passing -1 as new value)

+++ src/shardcache.c
@@ -2501,7 +2501,7 @@ int
 shardcache_arc_mode(shardcache_t *cache, arc_mode_t new_value)
 {
     int old_value = shardcache_get_set_option(&cache->arc_mode, (int)new_value);
-    if (new_value != -1 && old_value != new_value)
+    if ((int)new_value != -1 && old_value != new_value)
         arc_set_mode(cache->arc, new_value);
     return old_value;
 }
